Split sunbeam.cpp
Split file in order to enhance c++ source file structure
