+++ sunbeam/completion.hpp
@@ -1,0 +1,14 @@
+#ifndef SUNBEAM_COMPLETION_HPP
+#define SUNBEAM_COMPLETION_HPP
+
+#include <boost/python.hpp>
+
+
+namespace completion {
+    namespace py = boost::python;
+
+    void export_Completion();
+
+}
+
+#endif //SUNBEAM_COMPLETION_HPP
