+++ sunbeam/eclipse_config.hpp
@@ -1,0 +1,18 @@
+#ifndef SUNBEAM_ECLIPSE_CONFIG_HPP
+#define SUNBEAM_ECLIPSE_CONFIG_HPP
+
+#include <boost/python.hpp>
+
+
+namespace Opm {
+}
+
+namespace eclipse_config {
+    namespace py = boost::python;
+    using namespace Opm;
+
+    void export_EclipseConfig();
+
+}
+
+#endif //SUNBEAM_ECLIPSE_CONFIG_HPP
